requirements.txt
python-binance==1.0.17
python-dotenv==1.0.0
rich==13.7.0
pydantic==1.10.13
ccxt==4.1.13 